fix: 분리한 컴포넌트 rollback, 테스트 진행
