--- src/refactoring/components/cart/CartPageLayout.tsx
@@ -1,8 +1,0 @@
-export default function CartPageLayout({children}: { children: React.ReactNode }) {
-    return <div className="container mx-auto p-4">
-        <h1 className="text-3xl font-bold mb-6">장바구니</h1>
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-            {children}
-        </div>
-    </div>
-}
