+++ src/refactoring/hooks/useProduct.ts
@@ -2,9 +2,26 @@ import { useState } from "react";
 import { Product } from "../../types.ts";
 
 export const useProducts = (initialProducts: Product[]) => {
+  const [products, setProducts] = useState<Product[]>(initialProducts);
+
+  const updateProduct = (
+    productId: string,
+    updatedProduct: Partial<Product>,
+  ) => {
+    setProducts((prevProducts) =>
+      prevProducts.map((product) =>
+        product.id === productId ? { ...product, ...updatedProduct } : product,
+      ),
+    );
+  };
+
+  const addProduct = (newProduct: Product) => {
+    setProducts((prevProducts) => [...prevProducts, newProduct]);
+  };
+
   return {
-    products: [],
-    updateProduct: () => undefined,
-    addProduct: () => undefined,
+    products,
+    updateProduct,
+    addProduct,
   };
 };
