+++ src/refactoring/models/cart.ts
@@ -1,16 +1,20 @@
 import { CartItem, Coupon } from "../../types";
 
 export const calculateItemTotal = (item: CartItem) => {
-  return 0;
+  const basePrice = item.product.price * item.quantity;
+  return basePrice;
 };
 
 export const getMaxApplicableDiscount = (item: CartItem) => {
-  return 0;
+  const applicableDiscounts = item.product.discounts.filter(
+    (discount) => item.quantity >= discount.quantity,
+  );
+  return applicableDiscounts;
 };
 
 export const calculateCartTotal = (
   cart: CartItem[],
-  selectedCoupon: Coupon | null
+  selectedCoupon: Coupon | null,
 ) => {
   return {
     totalBeforeDiscount: 0,
@@ -22,7 +26,7 @@ export const calculateCartTotal = (
 export const updateCartItemQuantity = (
   cart: CartItem[],
   productId: string,
-  newQuantity: number
+  newQuantity: number,
 ): CartItem[] => {
   return [];
 };
